Relocated external content related tests from FedoraLdpIT to ExternalContentHandlerIT
